The repository of the list of cities is highlighted separately. Added Gradle GSON dependence.
