+++ app/src/main/AndroidManifest.xml
@@ -3,14 +3,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     package="ru.brauer.weather">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
     <application
         android:allowBackup="true"
+        android:debuggable="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Weather"
-        android:debuggable="true"
         tools:ignore="HardcodedDebugMode">
         <activity android:name=".ui.main.MainActivity">
             <intent-filter>
