+++ app/src/main/java/ru/brauer/weather/domain/repository/weather/loader/LoaderFromYandexWeather.kt
@@ -1,0 +1,5 @@
+package ru.brauer.weather.domain.repository.weather.loader
+
+class LoaderFromYandexWeather {
+
+}
